requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt --prerelease=allow --emit-find-links --refresh
--find-links https://storage.googleapis.com/jax-releases/libtpu_releases.html

absl-py==2.3.0
    # via
    #   -r requirements.in
    #   chex
    #   ml-collections
    #   optax
    #   orbax-checkpoint
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via gcsfs
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   anyioutils
    #   ipykernel
    #   zmq-anyio
anyioutils==0.7.3
    # via zmq-anyio
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via
    #   aiohttp
    #   outcome
cachetools==5.5.2
    # via google-auth
certifi==2025.6.15
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.2
    # via requests
chex==0.1.89
    # via optax
click==8.2.1
    # via wandb
comm==0.2.2
    # via ipykernel
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
debugpy==1.8.14
    # via ipykernel
decorator==5.2.1
    # via
    #   gcsfs
    #   ipython
docstring-parser==0.16
    # via tyro
einops==0.8.1
    # via -r requirements.in
etils==1.12.2
    # via orbax-checkpoint
executing==2.2.0
    # via stack-data
flax==0.10.6
    # via -r requirements.in
flax-orbax==0.1.4
    # via -r requirements.in
fonttools==4.58.4
    # via matplotlib
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.5.1
    # via
    #   etils
    #   gcsfs
gcsfs==2025.5.1
    # via -r requirements.in
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via wandb
google-api-core==2.25.1
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.40.3
    # via
    #   gcsfs
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
google-auth-oauthlib==1.2.2
    # via gcsfs
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-storage==3.1.0
    # via
    #   -r requirements.in
    #   gcsfs
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.70.0
    # via google-api-core
humanize==4.12.3
    # via orbax-checkpoint
idna==3.10
    # via
    #   anyio
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via opentelemetry-api
importlib-resources==6.5.2
    # via etils
ipykernel==7.0.0a1
    # via -r requirements.in
ipython==9.3.0
    # via ipykernel
ipython-pygments-lexers==1.1.1
    # via ipython
jax==0.6.1
    # via
    #   -r requirements.in
    #   chex
    #   flax
    #   optax
    #   orbax-checkpoint
jaxlib==0.6.1
    # via
    #   chex
    #   jax
    #   optax
jaxtyping==0.3.2
    # via -r requirements.in
jedi==0.19.2
    # via ipython
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.8.1
    # via
    #   ipykernel
    #   jupyter-client
kiwisolver==1.4.8
    # via matplotlib
libtpu==0.0.15
    # via jax
libtpu-nightly==0.1.dev20250616+nightly
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
matplotlib==3.10.3
    # via seaborn
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mdurl==0.1.2
    # via markdown-it-py
ml-collections==1.1.0
    # via -r requirements.in
ml-dtypes==0.5.1
    # via
    #   jax
    #   jaxlib
    #   tensorstore
msgpack==1.1.1
    # via
    #   flax
    #   orbax-checkpoint
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
nest-asyncio==1.6.0
    # via
    #   ipykernel
    #   orbax-checkpoint
numpy==1.26.4
    # via
    #   -r requirements.in
    #   chex
    #   contourpy
    #   flax
    #   jax
    #   jaxlib
    #   matplotlib
    #   ml-dtypes
    #   optax
    #   orbax-checkpoint
    #   pandas
    #   scipy
    #   seaborn
    #   tensorstore
    #   treescope
oauthlib==3.2.2
    # via requests-oauthlib
opentelemetry-api==1.34.1
    # via
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.34.1
    # via sentry-sdk
opentelemetry-semantic-conventions==0.55b1
    # via opentelemetry-sdk
opt-einsum==3.4.0
    # via jax
optax==0.2.5
    # via
    #   -r requirements.in
    #   flax
orbax-checkpoint==0.11.15
    # via
    #   -r requirements.in
    #   flax
outcome==1.3.0.post0
    # via anyioutils
packaging==25.0
    # via
    #   ipykernel
    #   matplotlib
    #   wandb
pandas==2.3.0
    # via
    #   -r requirements.in
    #   seaborn
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.2.1
    # via matplotlib
platformdirs==4.3.8
    # via
    #   jupyter-core
    #   wandb
prompt-toolkit==3.0.51
    # via ipython
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via google-api-core
protobuf==5.29.5
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   orbax-checkpoint
    #   proto-plus
    #   wandb
psutil==7.0.0
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via wandb
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.1
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   rich
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   flax
    #   flax-orbax
    #   ml-collections
    #   orbax-checkpoint
    #   wandb
pyzmq==26.4.0
    # via
    #   ipykernel
    #   jupyter-client
    #   zmq-anyio
requests==2.32.4
    # via
    #   gcsfs
    #   google-api-core
    #   google-cloud-storage
    #   jax
    #   requests-oauthlib
    #   wandb
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==14.0.0
    # via
    #   flax
    #   tyro
rsa==4.9.1
    # via google-auth
safetensors==0.6.0rc0
    # via -r requirements.in
scipy==1.16.0rc2
    # via
    #   jax
    #   jaxlib
seaborn==0.13.2
    # via -r requirements.in
sentry-sdk==3.0.0a2
    # via wandb
setuptools==80.9.0
    # via chex
shtab==1.7.2
    # via tyro
simplejson==3.20.1
    # via orbax-checkpoint
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
stack-data==0.6.3
    # via ipython
tensorstore==0.1.75
    # via
    #   flax
    #   orbax-checkpoint
toolz==1.0.0
    # via chex
tornado==6.5.1
    # via jupyter-client
tqdm==4.67.1
    # via -r requirements.in
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
treescope==0.1.9
    # via flax
typeguard==4.4.4
    # via tyro
typing-extensions==4.14.0
    # via
    #   anyio
    #   chex
    #   etils
    #   flax
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   orbax-checkpoint
    #   pydantic
    #   pydantic-core
    #   typeguard
    #   typing-inspection
    #   tyro
    #   wandb
typing-inspection==0.4.1
    # via pydantic
tyro==0.9.24
    # via -r requirements.in
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via
    #   requests
    #   sentry-sdk
wadler-lindig==0.1.6
    # via jaxtyping
wandb==0.20.2rc20250616
    # via -r requirements.in
wcwidth==0.2.13
    # via prompt-toolkit
yarl==1.20.1
    # via aiohttp
zipp==3.23.0
    # via
    #   etils
    #   importlib-metadata
zmq-anyio==0.3.9
    # via ipykernel
